requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv

faiss-cpu
pymongo
motor
httpx
numpy
scikit-learn
python-jose[cryptography]
passlib[bcrypt]
tenacity
structlog
rich
pytest
pytest-asyncio
